Fix simulation clock and refactor code

+++ src/physics/atmosphere.cpp
@@ -7,6 +7,6 @@ namespace Physics
 	float Atmosphere::airDensity(float height)
 	{
 		static constexpr float zeroAirDensityHeight = 44300;
-		return std::max(seaLevelAirDensity * (1 - height/zeroAirDensityHeight), (float)0);
+		return std::max(seaLevelAirDensity * (1 - height/zeroAirDensityHeight), 0.0f);
 	}
 };
